Removed comments, replaced Codable protocols
