+++ WeatherApp/APP/WeatherAppApp.swift
@@ -11,13 +11,11 @@ import SwiftData
 @main
 struct WeatherAppApp: App {
     @StateObject var viewModel = WeatherViewModel()
-    @StateObject var locationManager = LocationManager()
     
     var body: some Scene {
         WindowGroup {
             WeatherView()
                 .environmentObject(viewModel)
-                .environmentObject(locationManager)
                 .environment(\.modelContext, appContainer.mainContext)
             
         }
