+++ WeatherApp/Extensions/View.swift
@@ -8,7 +8,6 @@
 import SwiftUI
 
 extension View {
-    
     var screenWidth: CGFloat {
         UIScreen.main.bounds.width
     }
