+++ WeatherApp/Scene/Background Views/CloudyView.swift
